requirements.txt
pywin32>=306
pyautogui>=0.9.54
python-dotenv>=1.0.0
PySide6>=6.6.0
requests>=2.31.0
validators>=0.22.0
groq>=0.4.1